Refactor controllers, abstracted movement
