+++ Assets/Scripts/Controllers/MovementController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 527fddca998e05b41a35ee14bb001c64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
